+++ runtime/providers/postgres/pg.ts
@@ -7,7 +7,11 @@ import {
 } from "./adapter"
 
 export const provider = createPostgresProvider({
-  createPool: (connectionString) => new pg.Pool({ connectionString }),
+  createPool: (config) =>
+    new pg.Pool({
+      connectionString: config.connectionString,
+      max: config.maxConnections,
+    }),
   openCursor(client: DriverClient, query: DriverQuery) {
     return (client as pg.PoolClient).query(
       new Cursor(query.text, query.values as unknown[])